+add feature list view to home
