+++ lib/features/home/presentation/view/widget/custom_app_bar.dart
@@ -8,7 +8,7 @@ class CustomAppBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 30,vertical: 40),
+      padding: const EdgeInsets.only(left: 24,right: 24,top: 40,bottom: 20),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
