+++ lib/features/home/presentation/view/widget/home_view_body.dart
@@ -1,5 +1,6 @@
 import 'package:bookly_app/features/home/presentation/view/widget/custom_app_bar.dart';
-import 'package:bookly_app/features/home/presentation/view/widget/custom_list_view_item.dart';
+import 'package:bookly_app/features/home/presentation/view/widget/feature_list_view_item.dart';
+import 'package:bookly_app/features/home/presentation/view/widget/feature_books_list_view.dart';
 import 'package:flutter/material.dart';
 
 class HomeViewBody extends StatelessWidget {
@@ -10,7 +11,7 @@ class HomeViewBody extends StatelessWidget {
     return const Column(
       children: [
         CustomAppBar(),
-        CustomListViewItem(),
+        FeatureBooksListView(),
       ],
     );
   }
